Renamed refactor folder into tests
